% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_dolt_installed}
\alias{is_dolt_installed}
\alias{dolt_version}
\alias{dolt_path}
\title{Find and check for the presence of a dolt binary.}
\usage{
is_dolt_installed()

dolt_version()

dolt_path()
}
\description{
Find and check for the presence of a dolt binary.
}
