% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dolt}
\alias{dolt}
\alias{dbConnect,DoltDriver-method}
\alias{dbGetInfo,DoltConnection-method}
\alias{show,DoltConnection-method}
\title{Connect to a dolt database}
\usage{
DBI::dbConnect(doltr::dolt(), ...)

\S4method{dbConnect}{DoltDriver}(drv, dbname, username, password, host, port, ...)

\S4method{dbGetInfo}{DoltConnection}(dbObj, ...)

\S4method{show}{DoltConnection}(object)
}
\description{
\code{dolt()} is a DBI Driver for a Dolt Database. It, \code{DoltDriver},
and \code{DoltConnection} class are wrappers around the around classes and methods
from the \code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB}} package.
}
\details{
Methods not shown here fall back to those for \link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB}.
}
