% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dolt-config}
\alias{dolt-config}
\alias{dolt_config}
\alias{config}
\title{Configuration options}
\description{
The dolt package's behavior can be modified by setting these environment
variables.
}
\details{
\itemize{
\item \code{DOLT_PORT} sets the port to connect to or to run the server on. Defaults
to 3306. - \code{DOLT_HOST} sets the host IP to connect to or to run the server
on. Defaults to 127.0.0.1.
\item \code{DOLT_DIR} set the default directory to look for
a dolt database and run a server when using \code{\link[=dolt_local]{dolt_local()}} and \code{\link[=dolt]{dolt()}}.
You likely want to
set this on a project-level and .gitignore it when your project contains its
own database. Defaults to "doltdb".
\item \code{DOLT_CONFIG_FILE} is the path to a
file with additional configuration options for the dolt sql server. See
\url{https://docs.dolthub.com/interfaces/cli#dolt-sql-server} for options. -
\code{DOLT_PATH} specifies the path to the dolt binary if running locally.
Defaults to the one found in the system path.
\item \code{DOLT_COLLECT} specifies whether dolt
convenience functions returning data should return fully collected
tibbles or lazy tibbles for further processing.
\item \code{DOLT_SHOW_SQL} will print the SQL statements executed when running dolt
convenience functions.
}
}
