% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-diffs.R
\name{dolt_diffs}
\alias{dolt_diffs}
\alias{dolt_table_history}
\title{Examine information about dolt tables and diffs}
\usage{
dolt_diffs(table, to, from, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_table_history(table, conn = dolt(), collect = NULL, show_sql = NULL)
}
\arguments{
\item{table}{\link{character} the name of a table in the database}

\item{to}{commit to compare to}

\item{from}{commit to compare from}

\item{conn}{the database connection}

\item{collect}{whether to collect the result into R or return a \code{\link[dbplyr:tbl_lazy]{dbplyr::tbl_lazy()}}
to be further processed before collecting.  Defaults to \code{TRUE}, can be set with the \link[=dolt_vars]{environment variable}
\code{DOLT_COLLECT}.}

\item{show_sql}{Whether to print the SQL statements used internally to fetch
the data. Useful for learning how Dolt works internally. Defaults to \code{FALSE}, can
be set with the environment variable \code{DOLT_VERBOSE}.}
}
\description{
Examine information about dolt tables and diffs
}
