% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt-info}
\alias{dolt-info}
\alias{dolt_branches}
\alias{dolt_remotes}
\alias{dolt_docs}
\alias{dolt_log}
\title{Get info about the dolt database}
\usage{
dolt_branches(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_remotes(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_docs(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_log(conn = dolt(), collect = NULL, show_sql = NULL)
}
\description{
Get info about the dolt database
}
