% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local _connection.R
\name{dbConnect,DoltLocalDriver-method}
\alias{dbConnect,DoltLocalDriver-method}
\alias{doltdb}
\alias{dbGetInfo,DoltLocalConnection-method}
\alias{show,DoltLocalConnection-method}
\alias{dbDisconnect,DoltLocalConnection-method}
\title{Connect to a local dolt database directory}
\usage{
\S4method{dbConnect}{DoltLocalDriver}(
  drv,
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  port = Sys.getenv("DOLT_PORT", 3306L),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  find_server = TRUE,
  find_port = TRUE,
  ...
)

doltdb()

\S4method{dbGetInfo}{DoltLocalConnection}(dbObj, ...)

\S4method{show}{DoltLocalConnection}(object)

\S4method{dbDisconnect}{DoltLocalConnection}(conn, ...)
}
\arguments{
\item{find_server}{if TRUE, look for an existing dolt server process serving
the directory to connect to before starting a new server}

\item{find_port}{if TRUE, switch to a different port if \code{port} is used by
another process}

\item{...}{additional arguments to pass to \code{\link[=dolt_server]{dolt_server()}} to configure the
server}
}
\description{
\code{dolt_local()} creates a \code{DoltLocalDriver}, which can generate
a \code{DoltLocalConnection}.  Unlike \code{\link[=dolt]{dolt()}} and \code{DoltDriver}, this version
takes a directory name for an on-disk dolt database, starts and manages a
\link[=dolt_server]{dolt SQL server} in the background serving that directory,
connects to it and returns the connection. Parameters govern both the server
and connection

\code{doltdb(...)} is a convenience function for \verb{dbConnect(dolt_local),...)}

Connections to the same server, and servers of the same directory, are cached
and re-used, so re-connecting to the same directory is fast and
\code{doltdb()} can be invoked repeatedly and in simultaneous sessions, with little
overhead. Multi-user or other, more complicated networking steups should
use \code{dolt()} and \code{dolt_server()} directly.
}
