% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-methods.R
\name{dbSendQuery,DoltConnection,character-method}
\alias{dbSendQuery,DoltConnection,character-method}
\alias{dbSendStatement,DoltConnection,character-method}
\alias{dbClearResult,DoltResult-method}
\alias{dbDisconnect,DoltConnection-method}
\title{Miscellaneous Dolt Methods}
\usage{
\S4method{dbSendQuery}{DoltConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbSendStatement}{DoltConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbClearResult}{DoltResult}(res, ...)

\S4method{dbDisconnect}{DoltConnection}(conn, ...)
}
\arguments{
\item{conn}{an \linkS4class{DoltConnection} object.}

\item{statement}{a character vector of length one specifying the SQL
statement that should be executed.  Only a single SQL statement should be
provided.}

\item{params}{A list of query parameters to be substituted into
a parameterized query.}

\item{...}{Unused. Needed for compatibility with generic.#' @export}

\item{res}{A  \linkS4class{DoltResult} object.}
}
\description{
These methods largely wrap \code{RMariaDB} methods with small tweaks to work with
Dolt databases.
}
