% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_checkout}
\alias{dolt_checkout}
\alias{dolt_use}
\title{Navigate dolt history}
\usage{
dolt_checkout(
  ref,
  b = FALSE,
  start_point = NULL,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_use(hash = NULL, conn = dolt())
}
\arguments{
\item{b}{whether to create a new branch}

\item{hash}{the commit hash you want to set the database to. If NULL, checks
out the head of the main branch and brings you out of read-only mode.}

\item{branch}{the branch to check out}
}
\description{
\code{dolt_checkout()} checks out a dolt branch setting that branch as HEAD and
bringing you to it's tip. \code{dolt_use()} sets the database to use a specific
commit as it's state and puts you in read-only mode.
}
