% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-table.R
\name{dolt-read}
\alias{dolt-read}
\alias{dbReadTable,DoltConnection,character-method}
\alias{dbListTables,DoltConnection-method}
\alias{dbListObjects,DoltConnection-method}
\alias{dbExistsTable,DoltConnection,character-method}
\title{Read and write MariaDB tables.}
\usage{
\S4method{dbReadTable}{DoltConnection,character}(
  conn,
  name,
  as_of = NULL,
  ...,
  row.names = FALSE,
  check.names = TRUE
)

\S4method{dbListTables}{DoltConnection}(conn, as_of = NULL, ...)

\S4method{dbListObjects}{DoltConnection}(conn, prefix = NULL, as_of = NULL, ...)

\S4method{dbExistsTable}{DoltConnection,character}(conn, name, as_of = NULL, ...)
}
\arguments{
\item{conn}{a \linkS4class{DoltConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}} or \code{\link[=dolt]{dolt()}}}

\item{name}{a character string specifying a table name.}

\item{as_of}{A dolt commit hash, branch name, or object coercible to POSIXct}

\item{...}{Unused, needed for compatibility with generic.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{check.names}{If \code{TRUE}, the default, column names will be
converted to valid R identifiers.}

\item{prefix}{A fully qualified path in the database's namespace, or \code{NULL}.
This argument will be processed with \code{\link[DBI:dbUnquoteIdentifier]{dbUnquoteIdentifier()}}.
If given the method will return all objects accessible through this prefix.}
}
\value{
A data.frame in the case of \code{dbReadTable()}; a character vector of
names for \code{dbListTables()} and \code{dbListObjects()}, and a logical result for
\code{dbExistsTable()}.
}
\description{
These methods are extensions of standard DBI functions such as \link[DBI:dbReadTable]{DBI::dbReadTable}.
They differ in that they can take an \code{as_of} argument, reading historical data
from the database that was written as of a certain date or commit hash, or
from a different branch.
}
\seealso{
\href{https://www.dolthub.com/blog/2020-03-20-querying-historical-data-with-as-of/}{Querying Historical Data with AS OF Queries}
on the DoltHub blog, and \link[RMariaDB:mariadb-tables]{RMariaDB methods} upon
which these are built.
}
