% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_head_ref}
\alias{dolt_head_ref}
\alias{dolt_head}
\alias{dolt_head_hash}
\alias{dolt_staged_hash}
\alias{dolt_working_hash}
\title{Return the hashes and refspecs of the dolt database}
\usage{
dolt_head_ref(conn = dolt(), show_sql = NULL)

dolt_head(conn = dolt(), show_sql = NULL)

dolt_head_hash(conn = dolt(), show_sql = NULL)

dolt_staged_hash(conn = dolt(), show_sql = NULL)

dolt_working_hash(conn = dolt(), show_sql = NULL)
}
\arguments{
\item{conn}{The dolt database to query}

\item{show_sql}{Whether to print to the console the SQL statement executed.
Useful in learning dolt SQL syntax.  Defaults to false. Can be set globally
with the environment variable \code{\link[=dolt-config]{DOLT_SHOW_SQL}}.}
}
\description{
Return the hashes and refspecs of the dolt database
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_add}()},
\code{\link{dolt_diffs}()},
\code{\link{dolt_push}()},
\code{\link{dolt_status}()}
}
\concept{dolt-sql-commands}
