% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-stage-commit.R
\name{dolt_add}
\alias{dolt_add}
\alias{dolt_commit}
\alias{dolt_reset}
\title{Add, commit, and reset tables in a dolt database}
\usage{
dolt_add(tables = NULL, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_commit(
  all = TRUE,
  message = NULL,
  author = NULL,
  date = NULL,
  allow_empty = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_reset(
  hard = FALSE,
  tables = NULL,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)
}
\arguments{
\item{tables}{Which tables to be reset? Defaults to all tables if NULL.}

\item{conn}{the database connection}

\item{collect}{whether to collect the result into R or return a \code{\link[dbplyr:tbl_lazy]{dbplyr::tbl_lazy()}}
to be further processed before collecting.  Defaults to \code{TRUE}, can be set with the \link[=dolt_vars]{environment variable}
\code{DOLT_COLLECT}.}

\item{show_sql}{Whether to print the SQL statements used internally to fetch
the data. Useful for learning how Dolt works internally. Defaults to \code{FALSE}, can
be set with the environment variable \code{DOLT_VERBOSE}.}

\item{all}{stage all tables before committing?}

\item{message}{A commit message. If NULL in an interactive session, the user
will be prompted. Otherwise will error if empty.}

\item{author, date}{Author and date. If null, uses the ones set in \link{dolt-config}.
Author should be in the format \code{"A U Thor author@example.com"}}

\item{allow_empty}{Allow recording a commit that has the exact same data as
its sole parent. This is usually a mistake, so it is FALSE by default.}

\item{hard}{Reset working and staged tables? If FALSE (default), a "soft"
reset will be performed, only unstaging staged tables.  If TRUE, all
working and staged changes will be discarded.}
}
\description{
Add, commit, and reset tables in a dolt database
}
