% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{dolt_state}
\alias{dolt_state}
\alias{dolt_status}
\alias{dolt_last_commit}
\title{Get information about a dolt database}
\usage{
dolt_state(conn = dolt())

dolt_status(conn = dolt())

dolt_last_commit(conn = dolt())
}
\value{
A data frame of class "dolt_status" and \link[tibble:tbl_df-class]{tibble::tbl_df}.  It
pretty-prints as an abbreviated summary of status.
}
\description{
These functions yield information about the current state of a dolt database.
\code{dolt_state()} provides information on current branch or headless commit.
\code{dolt_status()} summarizes changes to the database in working or staged tables (from the \code{dolt_status} table).
\code{dolt_last_commit()} pulls the most recent value from the \code{dolt_log} table.  All
have pretty-print methods for the objects returned but can be interrogated for more detail.
}
\details{
Values from each of these functions are returned as part of the \code{dbGetInfo()}
method and are part of the information shown in the \code{DoltConnection} print
method and in the RStudio Connection pane for a Dolt Database.
}
