% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-types.R
\name{dbDataType,DoltConnection-method}
\alias{dbDataType,DoltConnection-method}
\alias{dolt_type_sizes}
\title{Dolt Data Types}
\usage{
\S4method{dbDataType}{DoltConnection}(
  dbObj,
  obj,
  min_varchar = Sys.getenv("DOLT_MINVARCAHR", 255L),
  max_varchar = Sys.getenv("DOLT_MAXVARCHAR", 16383L),
  ...
)

dolt_type_sizes(types)
}
\arguments{
\item{dbObj}{the database connection}

\item{obj}{the data type (vector or data frame)}

\item{min_varchar}{The minimum size \code{VARCHAR} types should be cast as}

\item{max_varchar}{the maximum size \code{VARCHAR} types should be cast as. Larger
text data will return types \code{TEXT},\code{MEDIUMTEXT}, or \code{LONGTEXT}}

\item{...}{further arguments to methods}

\item{types}{a character vector of dolt types, e.g., \code{"VARCHAR(12)"},
\code{"LONGBLOB"}, \code{"TINYINT"}, etc.}
}
\value{
A character vector of classes, with attributes of the maximum size for
text and blob classes
}
\description{
\code{dbDataType} matches R data types to dolt data types. For text and blob data
types, it automatically chooses amongst options (e.g., \verb{VARCHAR(N), }TEXT\verb{, }LONGTEXT\verb{, etc.}) based on maximum field length.  An attribute of maximum
size of these fields is returned to support operations where fields need to be recast in revision.
}
\details{
\code{dolt_type_sizes()} takes a vector of SQL types and returns the maximum field
size, if applicable.
}
