% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-vars.R
\name{dolt-vars}
\alias{dolt-vars}
\alias{dolt_vars}
\alias{env_vars}
\alias{environment_variables}
\title{Configuration variable options}
\description{
The doltr package's behavior can be modified by setting these environment
variables:
}
\details{
\itemize{
\item \code{DOLT_DIR} set the default directory to look for
a dolt database and run a server when using \code{\link[=dolt_local]{dolt_local()}} and \code{\link[=dolt]{dolt()}}.
Defaults to "doltdb".
\item \code{DOLT_PORT} sets the port to connect to or to run the server on. Defaults
to 3306.
\item \code{DOLT_HOST} sets the host IP to connect to or to run the server
on. Defaults to 127.0.0.1.
\item \code{DOLT_CONFIG_FILE} is the path to a
file with additional configuration options for the dolt sql server. See
\url{https://docs.dolthub.com/interfaces/cli#dolt-sql-server} for options.
\item \code{DOLT_PATH} specifies the path to the dolt binary if running locally.
Defaults to the one found in the system path.
\item \code{DOLT_COLLECT} specifies whether dolt
convenience functions returning data should return fully collected
tibbles or lazy tibbles for further processing. Set it to 0 or \code{false}
to disable, potentially for when large databases with long histories yield
very large responses to commands like \code{dolt_log()} or \code{dolt_diffs()}.
\item \code{DOLT_VERBOSE} will print the SQL or command-line statements executed when
running functions that wrap database or system calls. Useful for
understanding how dolt commands work. Set to 1 or \code{true} to enable this
behavior.
\item \code{DOLT_WATCH} determines whether the RStudio Connection pane automatically
updates in response to changes in the database.  Set it to 0 or \code{false}
to disable this behavior.
\item \code{DOLT_ROOT_DIR} the directory where Dolt global configuration and credential
data is stored (\verb{~/.dolt} by default). Note this can also be set in your
shell to configure command-line dolt.
}
}
\seealso{
dolt-config
}
