% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-table.R
\name{dbWriteTable,DoltConnection,character,data.frame-method}
\alias{dbWriteTable,DoltConnection,character,data.frame-method}
\title{Write a table to the database}
\usage{
\S4method{dbWriteTable}{DoltConnection,character,data.frame}(
  conn,
  name,
  value,
  field.types = NULL,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  temporary = FALSE,
  batch_size = NULL,
  ...
)
}
\arguments{
\item{conn}{a database connection}

\item{name}{the table name}

\item{value}{A data frame.}

\item{field.types}{Optional, overrides default choices of field types,
derived from the classes of the columns in the data frame. See \link[=dbDataType,DoltConnection-method]{dbDataType()}}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{overwrite}{a logical specifying whether to overwrite an existing table
or not. Its default is \code{FALSE}.}

\item{append}{a logical specifying whether to append to an existing table
in the database  If appending, then the table (or temporary table)
must exist, otherwise an error is reported. Its default is \code{FALSE}.}

\item{temporary}{If \code{TRUE}, creates a temporary table that expires
when the connection is closed. For \code{dbRemoveTable()}, only temporary
tables are considered if this argument is set to \code{TRUE}}

\item{batch_size}{The number of records to insert in a single statement (defaults to all)}

\item{...}{for additional parameters passed on. Not currently used.}
}
\description{
This method uses \code{\link[dbx:dbxInsert]{dbx::dbxInsert()}} as that implementation
is much more performant than the standard method from \link[RMariaDB:mariadb-tables]{RMariaDB::dbWriteTable()},
due to the way Dolt handles repeat \code{INSERT} statements.
}
\details{
This the dependency on \code{dbx} may be removed if the base issue is resolved: \url{https://github.com/dolthub/dolt/issues/2091}.
}
\seealso{
dolt-read
}
