% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt.R
\name{dolt}
\alias{dolt}
\title{Return a (cached) connection to the default Dolt database}
\usage{
dolt(
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  dbname = NULL,
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  port = Sys.getenv("DOLT_PORT", 3306L),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  cache_connection = TRUE,
  ...
)
}
\arguments{
\item{dir}{The directory from which to server a \code{\link[=dolt_local]{dolt_local()}} connection.
If \code{"remote"} a \code{\link[=dolt_remote]{dolt_remote()}} connection will be made and no server will
be started.}

\item{username}{The username. Defaults to "root"}

\item{password}{The login password.  Defaults to empty.}

\item{port}{The TCP port for connections. Defaults to 3306.}

\item{host}{The IP of the host. Defaults to the local machine, \verb{127.0.0.1}}

\item{cache_connection}{Should we preserve a cache of the connection? allows
faster load times and prevents connection from being garbage-collected.}

\item{...}{further arguments passed to \code{\link[=dolt_server]{dolt_server()}} or \code{\link[=MariaDB]{MariaDB()}}}
}
\description{
\code{dolt()} returns a connection to a default database. It is a convenience
wrapper around \verb{dbConnect(dolt_local/remote(), ...} that also caches connections
for faster loading.
}
\seealso{
Other connections: 
\code{\link{dolt_local}()},
\code{\link{dolt_remote}()}
}
\concept{connections}
