% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{dolt_clone}
\alias{dolt_clone}
\title{Clone a dolt database}
\usage{
dolt_clone(
  remote_url,
  remote = "origin",
  new_dir = basename(remote),
  branch = NULL
)
}
\value{
The directory of the cloned database (invisibly)
}
\description{
Clone a dolt database
}
