% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{dolt_dump}
\alias{dolt_dump}
\title{Export data from a dolt database}
\usage{
dolt_dump(
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  format = c("sql", "csv", "json"),
  out = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dir}{path to the directory. Will be created if it does not exist}

\item{format}{the export data format. One of \code{"sql"}, \code{"csv"}, or \code{"json"}}

\item{out}{the location on-disk for export. In the case of \code{"sql"}, format,
a single file path (default \code{doltdump.sql}), otherwise a directory for all
tables to be dumped as separate files (default "doltdump")}

\item{overwrite}{whether to overwrite existing files/directories.}
}
\value{
the path(s) of exported files
}
\description{
Export data from a dolt database
}
