% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{dolt_init}
\alias{dolt_init}
\title{Initiate a dolt database directory}
\usage{
dolt_init(dir = Sys.getenv("DOLT_DIR", "doltdb"))
}
\arguments{
\item{dir}{path to the directory. Will be created if it does not exist}
}
\description{
Initiate a dolt database directory
}
