% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{dolt_pane}
\alias{dolt_pane}
\title{Open a Dolt connection pane in RStudio}
\usage{
dolt_pane(conn = doltr::dolt())
}
\value{
The connection object (invisibly)
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) dolt_pane()
}
