% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-remote.R
\name{dolt_remote}
\alias{dolt_remote}
\alias{dbConnect,DoltDriver-method}
\title{Connect to a dolt database}
\usage{
DBI::dbConnect(doltr::dolt_remote(), ...)

\S4method{dbConnect}{DoltDriver}(
  drv,
  dbname = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  port = Sys.getenv("DOLT_PORT", 3306L),
  ...
)
}
\arguments{
\item{drv}{an object of class \code{DoltDriver}, created by \code{dolt_remote()}.}

\item{dbname}{The database name}

\item{username}{The username. Defaults to "root"}

\item{password}{The login password.  Defaults to empty.}

\item{host}{The IP of the host. Defaults to the local machine, \verb{127.0.0.1}}

\item{port}{The TCP port for connections. Defaults to 3306.}

\item{...}{other arguments passed to \link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB::MariaDB}}
}
\description{
\code{dolt_remote()} is a DBI Driver to connect to a dolt server via
a port. It, \code{DoltDriver},
and \code{DoltConnection} class are wrappers around the around classes and methods
from the \code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB}} package.

Most parameters can be specified with environment variables See \link{config}.
}
\details{
Most methods fall back to those for \code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB}}.
}
\seealso{
Other connections: 
\code{\link{dolt_local}()},
\code{\link{dolt}()}
}
\concept{connections}
