% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{dolt_server}
\alias{dolt_server}
\title{Fire up a dolt server and return the process ID of the shell process
If already running, return the PID of the process}
\usage{
dolt_server(
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  port = Sys.getenv("DOLT_PORT", 3306L),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  multi_db = FALSE,
  autocommit = TRUE,
  read_only = FALSE,
  log_level = "info",
  log_out = FALSE,
  timeout = 0,
  query_parallelism = 2,
  max_connections = 100,
  config_file = NULL
)
}
\description{
Fire up a dolt server and return the process ID of the shell process
If already running, return the PID of the process
}
