% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{doltHasDefault}
\alias{doltHasDefault}
\alias{doltDefault}
\title{Check if default dolt database is available.}
\usage{
doltHasDefault()

doltDefault()
}
\description{
RMariaDB examples and tests connect to a database defined by the
\code{rs-dbi} group in \verb{~/.my.cnf}. This function checks if that
database is available, and if not, displays an informative message.
\code{mariadbDefault()} works similarly but throws a testthat skip condition
on failure, making it suitable for use in tests.
}
\examples{
if (doltHasDefault()) {
  db <- dbConnect(doltr::dolt_local())
  dbListTables(db)
  dbDisconnect(db)
}
}
