% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doltr-package.R
\docType{package}
\name{doltr-package}
\alias{doltr}
\alias{doltr-package}
\title{doltr: A client for the dolt database}
\description{
Creates a DBI-compliant interface to dolt databases (<https://www.dolthub.com>). Also manages local dolt server processes, provides convenience functions for dolt versioning, and an RStudio connection pane interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ecohealthalliance.github.io/doltr}
  \item \url{https://github.com/ecohealthalliance/doltr}
  \item Report bugs at \url{https://github.com/ecohealthalliance/doltr/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{ross@ecohealthalliance.org} (\href{https://orcid.org/0000-0002-2136-0000}{ORCID})

Other contributors:
\itemize{
  \item EcoHealth Alliance [copyright holder]
}

}
\keyword{internal}
