% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-remote.R
\name{dbDataType,DoltConnection-method}
\alias{dbDataType,DoltConnection-method}
\alias{dolt_type_sizes}
\title{Dolt Data Types}
\usage{
\S4method{dbDataType}{DoltConnection}(
  dbObj,
  obj,
  min_varchar = Sys.getenv("DOLT_MINVARCAHR", 255L),
  max_varchar = Sys.getenv("DOLT_MAXVARCHAR", 16383L),
  ...
)

dolt_type_sizes(types)
}
\arguments{
\item{dbObj}{the database connection}

\item{obj}{the data type (vector or data frame)}

\item{min_varchar}{The minimum size \code{VARCHAR} types should be cast as}

\item{max_varchar}{the maximum size \code{VARCHAR} types should be cast as. Larger
text data will return types \code{TEXT},\code{MEDIUMTEXT}, or \code{LONGTEXT}}
}
\value{
A character vector of classes, with attributes of the maximum size for
text and blob classes
}
\description{
\code{dbDataType} fits more precise data types and returns a data type size attribute
in case fields need to be recast in revision. \code{dolt_type_sizes}
}
