% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-config.R
\name{dolt-config}
\alias{dolt-config}
\alias{dolt_config_get}
\alias{dolt_config_set}
\title{Get and set Dolt configuration variables}
\usage{
dolt_config_get(
  params = NULL,
  global = TRUE,
  local_dir = Sys.getenv("DOLT_DIR")
)

dolt_config_set(params, global = TRUE, local_dir = Sys.getenv("DOLT_DIR"))
}
\arguments{
\item{params}{What parameters to get or set. Can include \code{user.name}, \code{user.email},
and \code{user.creds}.  For \code{dolt_config_set}, this should be a named character
vector or list with parameter names and values.}

\item{global}{Set global or database-specific credentials}

\item{local_dir}{if not \code{global}, what local database to set variables for}
}
\description{
Get and set Dolt configuration variables
}
\seealso{
dolt_vars
}
