% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_diffs}
\alias{dolt_diffs}
\alias{dolt_table_history}
\title{Examine information about dolt tables and diffs}
\usage{
dolt_diffs(table, to, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_table_history(table, conn = dolt(), collect = NULL, show_sql = NULL)
}
\arguments{
\item{table}{\link{character} the name of a table in the database}

\item{to}{commit to compare to}
}
\description{
Examine information about dolt tables and diffs
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_add}()},
\code{\link{dolt_push}()}
}
\concept{dolt-sql-commands}
