% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_diffs}
\alias{dolt_diffs}
\alias{dolt_table_history}
\title{Examine information about dolt tables and diffs}
\usage{
dolt_diffs(table, to, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_table_history(table, conn = dolt(), collect = NULL, show_sql = NULL)
}
\arguments{
\item{table}{\link{character} the name of a table in the database}

\item{to}{commit to compare to}

\item{conn}{The dolt database to query}

\item{collect}{Whether to return the full results (TRUE) or a an uncollected
\code{\link[dplyr:tbl]{dplyr::tbl()}} to be further processed by dplyr pipelines before collecting.
Can be set globally with the environment variable \code{\link[=dolt-config]{DOLT_COLLECT}}.}

\item{show_sql}{Whether to print to the console the SQL statement executed.
Useful in learning dolt SQL syntax.  Defaults to false. Can be set globally
with the environment variable \code{\link[=dolt-config]{DOLT_SHOW_SQL}}.}
}
\description{
Examine information about dolt tables and diffs
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_add}()},
\code{\link{dolt_head_ref}()},
\code{\link{dolt_push}()},
\code{\link{dolt_status}()}
}
\concept{dolt-sql-commands}
