% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_status}
\alias{dolt_status}
\alias{dolt_branches}
\alias{dolt_remotes}
\alias{dolt_docs}
\alias{dolt_log}
\title{Get information about the dolt database}
\usage{
dolt_status(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_branches(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_remotes(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_docs(conn = dolt(), collect = NULL, show_sql = NULL)

dolt_log(conn = dolt(), collect = NULL, show_sql = NULL)
}
\arguments{
\item{conn}{The dolt database to query}

\item{collect}{Whether to return the full results (TRUE) or a an uncollected
\code{\link[dplyr:tbl]{dplyr::tbl()}} to be further processed by dplyr pipelines before collecting.
Can be set globally with the environment variable \code{\link[=dolt-config]{DOLT_COLLECT}}.}

\item{show_sql}{Whether to print to the console the SQL statement executed.
Useful in learning dolt SQL syntax.  Defaults to false. Can be set globally
with the environment variable \code{\link[=dolt-config]{DOLT_SHOW_SQL}}.}
}
\description{
Get information about the dolt database
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_add}()},
\code{\link{dolt_diffs}()},
\code{\link{dolt_head_ref}()},
\code{\link{dolt_push}()}
}
\concept{dolt-sql-commands}
