% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-info.R
\name{dbGetInfo,DoltConnection-method}
\alias{dbGetInfo,DoltConnection-method}
\alias{show,DoltConnection-method}
\title{Get information about a Dolt Database}
\usage{
\S4method{dbGetInfo}{DoltConnection}(dbObj, ...)

\S4method{show}{DoltConnection}(object)
}
\arguments{
\item{dbObj}{the database connection}

\item{...}{Other arguments to methods}

\item{object}{the database connection}
}
\description{
The Dolt \code{dbGetInfo()} returns standard information about a database connection
according to the \link[DBI:dbGetInfo]{DBI specification}, as well as information
about the version-control status of the repository such as current branch, last
commit, and modified tables.  This information is
also displayed in the print method for a dolt connection object and in the
\link[=dolt_pane]{RStudio connection pane}.
}
\seealso{
dolt_state dolt_status dolt_last_commit dolt_pane
}
