% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-nav.R
\name{dolt_checkout}
\alias{dolt_checkout}
\alias{dolt_use}
\title{Navigate dolt history}
\usage{
dolt_checkout(
  branch,
  b = FALSE,
  start_point = NULL,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_use(hash = NULL, conn = dolt())
}
\arguments{
\item{branch}{the branch to check out}

\item{b}{whether to create a new branch}

\item{start_point}{a commit hash from which the branch should start. If NULL,
starts from current HEAD.}

\item{conn}{the database connection}

\item{collect}{whether to collect the result into R or return a \code{\link[dbplyr:tbl_lazy]{dbplyr::tbl_lazy()}}
to be further processed before collecting.  Defaults to \code{TRUE}, can be set with the \link[=dolt_vars]{environment variable}
\code{DOLT_COLLECT}.}

\item{show_sql}{Whether to print the SQL statements used internally to fetch
the data. Useful for learning how Dolt works internally. Defaults to \code{FALSE}, can
be set with the environment variable \code{DOLT_VERBOSE}.}

\item{hash}{the commit hash you want to set the database to. If NULL, checks
out the head of the main branch and brings you out of read-only mode.}
}
\description{
\code{dolt_checkout()} checks out a dolt branch, setting that branch as HEAD and
bringing you to its tip. \code{dolt_use()} sets the database to use a specific
commit as it's state and puts you in read-only mode.
}
