% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_push}
\alias{dolt_push}
\alias{dolt_pull}
\alias{dolt_fetch}
\title{Work with dolt repository remotes}
\usage{
dolt_push(
  remote = NULL,
  ref = NULL,
  set_upstream = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_pull(
  remote = NULL,
  squash = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_fetch(
  remote = NULL,
  ref = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)
}
\arguments{
\item{remote}{the name of the remote. "origin" is used by default}

\item{ref}{the branch reference}

\item{set_upstream}{whether to set the remote branch reference to track}

\item{force}{whether to overwrite any conflicting history
the current branch}

\item{squash}{whether to merge changes to the working set without
updating the commit history}
}
\description{
Work with dolt repository remotes
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_add}()},
\code{\link{dolt_diffs}()}
}
\concept{dolt-sql-commands}
