% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolt-remote.R
\name{dolt_push}
\alias{dolt_push}
\alias{dolt_pull}
\alias{dolt_fetch}
\alias{dolt_clone}
\title{Work with dolt repository remotes}
\usage{
dolt_push(
  remote = NULL,
  ref = NULL,
  set_upstream = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_pull(
  remote = NULL,
  squash = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_fetch(
  remote = NULL,
  ref = FALSE,
  force = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_clone(
  remote_url,
  remote = "origin",
  new_dir = basename(remote_url),
  branch = NULL
)
}
\arguments{
\item{remote}{the name of the remote. "origin" is used by default}

\item{ref}{the branch reference}

\item{set_upstream}{whether to set the remote branch reference to track}

\item{force}{whether to overwrite any conflicting history
the current branch}

\item{conn}{the database connection}

\item{collect}{whether to collect the result into R or return a \code{\link[dbplyr:tbl_lazy]{dbplyr::tbl_lazy()}}
to be further processed before collecting.  Defaults to \code{TRUE}, can be set with the \link[=dolt_vars]{environment variable}
\code{DOLT_COLLECT}.}

\item{show_sql}{Whether to print the SQL statements used internally to fetch
the data. Useful for learning how Dolt works internally. Defaults to \code{FALSE}, can
be set with the environment variable \code{DOLT_VERBOSE}.}

\item{squash}{whether to merge changes to the working set without
updating the commit history}

\item{remote_url}{the remote URL to clone}

\item{new_dir}{the directory to clone into}

\item{branch}{the branch to clone. If NULL, clones all branches}

\item{remote_branch}{the name of the branch when setting upstream branch. The current local branch is used by default}
}
\description{
Work with dolt repository remotes
}
\concept{dolt-sql-commands}
