% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{dolt_add}
\alias{dolt_add}
\alias{dolt_commit}
\alias{dolt_reset}
\title{Add, commit, and reset tables in a dolt database}
\usage{
dolt_add(tables = NULL, conn = dolt(), collect = NULL, show_sql = NULL)

dolt_commit(
  all = TRUE,
  message = NULL,
  author = NULL,
  date = NULL,
  allow_empty = FALSE,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)

dolt_reset(
  hard = FALSE,
  tables = NULL,
  conn = dolt(),
  collect = NULL,
  show_sql = NULL
)
}
\arguments{
\item{tables}{Which tables to be reset? Defaults to all tables if NULL.}

\item{all}{stage all tables before comitting?}

\item{message}{A commit message. If NULL in an interactive session, the user
will be prompted. Otherwise will error if empty.}

\item{hard}{Reset working and staged tables? If FALSE (default), a "soft"
reset, of only staged tables, will be performed}
}
\description{
Add, commit, and reset tables in a dolt database
}
\seealso{
Other dolt-sql-commands: 
\code{\link{dolt_diffs}()},
\code{\link{dolt_push}()}
}
\concept{dolt-sql-commands}
