% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-local.R
\name{dolt_local}
\alias{dolt_local}
\alias{dbConnect,DoltLocalDriver-method}
\title{Connect to a local dolt database directory}
\usage{
DBI::dbConnect(doltr::dolt_local(), ...)

\S4method{dbConnect}{DoltLocalDriver}(
  drv,
  dir = Sys.getenv("DOLT_DIR", "doltdb"),
  username = Sys.getenv("DOLT_USERNAME", "root"),
  password = Sys.getenv("DOLT_PASSWORD", ""),
  port = Sys.getenv("DOLT_PORT", 3306L),
  host = Sys.getenv("DOLT_HOST", "127.0.0.1"),
  find_port = TRUE,
  server_args = list(),
  ...
)
}
\arguments{
\item{drv}{an object of class \code{DoltLocalDriver}, created by \code{dolt_local()}.}

\item{dir}{The dolt directory to serve and connect to}

\item{username}{The username. Defaults to "root"}

\item{password}{The login password.  Defaults to empty.}

\item{port}{The TCP port for connections. Defaults to 3306.}

\item{host}{The IP of the host. Defaults to the local machine, \verb{127.0.0.1}}

\item{server_args}{a list of additional arguments to pass to \code{\link[=dolt_server]{dolt_server()}}}

\item{...}{additional arguments to pass to \code{\link[RMariaDB:dbConnect-MariaDBDriver-method]{RMariaDB}}}
}
\description{
\code{dolt_local()} creates a \code{DoltLocalDriver}, which can generate
a \code{DoltLocalConnection}.  Unlike \code{\link[=dolt_remote]{dolt_remote()}} and \code{DoltDriver}, this version
takes a directory name for an on-disk dolt database, starts and manages a
\link[=dolt_server]{dolt SQL server} in the background serving that directory,
connects to it and returns the connection. Parameters govern both the server
and connection

Multi-user or other, more complicated networking set-ups should
use \code{\link[=dolt_server]{dolt_server()}} and \code{\link[=dolt_remote]{dolt_remote()}} directly.
}
\seealso{
Other connections: 
\code{\link{dolt_remote}()},
\code{\link{dolt}()}
}
\concept{connections}
