% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pane.R
\name{dolt_pane}
\alias{dolt_pane}
\title{Open a Dolt connection pane in RStudio}
\usage{
dolt_pane(conn = doltr::dolt())
}
\arguments{
\item{conn}{a dolt connection. If a path is provided instead, a connection
will be created to the path using \code{\link[=dolt]{dolt()}}.}
}
\value{
The connection object (invisibly)
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.  The pane will show the database versioning state,
tables stored in the database, and dolt system tables showing history.
}
\details{
When running dolt interactively, the connection pane will automatically
update in respone to most queries that modify the database state.  You
can stop this behavior by setting the \code{DOLT_WATCH} environment variable
to \code{0} or \code{false}.  See \link{config} for more configuration varuables.
}
\examples{
if (!is.null(getOption("connectionObserver"))) dolt_pane()
}
