% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doltr-package.R
\docType{package}
\name{doltr-package}
\alias{doltr}
\alias{doltr-package}
\title{doltr: A client for the dolt database}
\description{
Creates a DBI-compliant interface to dolt databases (<https://www.dolthub.com>). Also manages local dolt server processes, provides convenience functions for dolt versioning, and an RStudio connection pane interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecohealthalliance/doltr}
  \item Report bugs at \url{https://github.com/ecohealthalliance/doltr/issues}
}

}
\author{
\strong{Maintainer}: Noam Ross \email{noam.ross@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
